+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- "Livi" video player in the list of predefined video players.
+
 ## [1.14.2] - 2023-12-08
 
 ### Added